refactor: add ServerInviteTable component
