+++ src/pages/invites/index.tsx
@@ -1,27 +1,14 @@
-import {
-  Card,
-  styled,
-  Table,
-  TableBody,
-  TableCell as MuiTableCell,
-  TableHead,
-  TableRow,
-  Typography,
-} from "@mui/material";
+import { Typography } from "@mui/material";
 import { NextPage } from "next";
 import { useTranslation } from "react-i18next";
 import { useServerInvitesQuery } from "../../apollo/gen";
-import ServerInviteRow from "../../components/ServerInvites/ServerInviteRow";
 import ServerInviteForm from "../../components/ServerInvites/ServerInviteForm";
+import ServerInviteTable from "../../components/ServerInvites/ServerInviteTable";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import ProgressBar from "../../components/Shared/ProgressBar";
 import { isDeniedAccess } from "../../utils/error.utils";
 
-const TableCell = styled(MuiTableCell)(({ theme }) => ({
-  color: theme.palette.text.secondary,
-}));
-
-const ServerRoles: NextPage = () => {
+const ServerInvites: NextPage = () => {
   const { data, loading, error } = useServerInvitesQuery();
   const { t } = useTranslation();
 
@@ -42,31 +29,7 @@ const ServerRoles: NextPage = () => {
       return null;
     }
     const { me, serverInvites } = data;
-    return (
-      <Card>
-        <Table>
-          <TableHead>
-            <TableRow>
-              <TableCell>{t("invites.columnNames.inviter")}</TableCell>
-              <TableCell>{t("invites.columnNames.code")}</TableCell>
-              <TableCell>{t("invites.columnNames.uses")}</TableCell>
-              <TableCell>{t("invites.columnNames.expires")}</TableCell>
-              <TableCell></TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {serverInvites.map((serverInvite, index) => (
-              <ServerInviteRow
-                isLast={index + 1 === serverInvites.length}
-                key={serverInvite.id}
-                me={me}
-                serverInvite={serverInvite}
-              />
-            ))}
-          </TableBody>
-        </Table>
-      </Card>
-    );
+    return <ServerInviteTable serverInvites={serverInvites} me={me} />;
   };
 
   return (
@@ -81,4 +44,4 @@ const ServerRoles: NextPage = () => {
   );
 };
 
-export default ServerRoles;
+export default ServerInvites;
